perf: Reduce number of tensor allocations in `scale_local`
Move creation of `point_indeces` out of the loop that iterates over the
batches to only allocated the memory once.
It is then zeroed after each iteration.
This change should yield huge performance improvements over large batch
sizes.

This is possible because `num_items` and `num_features` is constant over
all the batches, so the shapes stay the same.