build: Update python module to C++20
NOTE: This only works for GCC currently, since it appears that there is
no portable way to set this.
This is why I'm currently working on completely moving to CMake for the
C++ builds.

But it should be fine for now, since this is only used on Linux anymay
(at the moment).