refactor: Pre-allocate precision instead of copying `true_positive`
Create and resize vector to avoid copy.
Then emplace values in for-loop.
I could've done this with std::generate as well but decided against it
because it would've involved increasing iterators manually and saving
the value in a temporary variable.
Therefore it is not clear whether that would've been more efficient, and
it definitely would've been less readable.