build(module): Fix build on Linux
It seems that `cnpy` does not include include `zlib` in the statically
linked library, because that would require a statically linked `zlib`,
which is not available on most platform.

So as a hotfix `zlib` will be linked when building the module as well
until `cnpy` maybe has a release build with a statically linked `zlib`
that one can pull directly.