fix: Fix `TraceTest`
Fix OpenMP setup fixing `TraceTest`.

For some reason the `omp parallel` region does not work the same way in a
mixed Python-C++ environment as it does in a pure C++ environment.
So the setup that worked for the mixed environment caused a duplication
of work in the C++ environment.

This is fixed by wrapping everything in a `omp parallel` followed by an
`omp single` so that OpenMP is still properly initialized in a mixed
environment but does not duplicate work in a pure C++ environment (hence
failing the tests).